set res.status

+++ src/controllers/pacienteController.js
@@ -10,11 +10,16 @@ const pacientesController = {
     //Listar Paciente
     listarPacientesId: async (req, res) => {
         const { id } = req.params;
-        const listaPacientes = await Pacientes.findByPk (id);
+        const listaPacientes = await Pacientes.findByPk(id);
+        if (!listaPacientes){
+            res.status(404).json("Id não encontrado");
+            return;
+        }
         res.status(200).json(listaPacientes);
     },
     //cadastro de paciente
     async cadastrarPaciente(req, res) {
+        try{
         const { nome, email, idade } = req.body;
         const pacienteNovo = await Pacientes.create({
             nome,
@@ -22,12 +27,20 @@ const pacientesController = {
             idade,
             
         });
-        res.status(201).json(pacienteNovo);
+        res.status(201).json(pacienteNovo);}
+        catch(error){ 
+            res.status(400).json("Erro na requisição. Todos os campos são obrigatórios");
+        }
     },
     //deletar paciente
     async deletePaciente(req, res) {
         try{
             const { id } = req.params;
+            const Procurardel = await Pacientes.findByPk(id);
+            if (!Procurardel){
+                res.status(404).json("Id não encontrado");
+                return;
+            }
             await Pacientes.destroy({
                 where: { id },
             });
@@ -39,6 +52,7 @@ const pacientesController = {
     },
     //atualizar paciente
     async atualizarPaciente(req, res) {
+        try{
         const {id} = req.params;
         const {nome, email, idade,} = req.body;
         const pacienteNovo = await Pacientes.update({
@@ -52,7 +66,10 @@ const pacientesController = {
                 },
             });
         res.status(200).json("Paciente atualizado");
-        
+        }
+        catch(error){
+            res.status(400).json("Erro na requisição. Todos os campos são obrigatórios");
+        }
     }
 };
